Add retry loop to Launcher
Sometimes we can get an authentication error while trying to login to a
new node if connecting too quickly. Our systems do not create a jenkins
user account until the system has been fully commissioned so it could
take a 10s of seconds before authetication is possible.

This commit adds a retry loop to the launcher so that we do not assume
the system is not available.

Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>