WIP | Fix for JENKINS-42207
The OpenStack plugin fails to wait for JNLP connections. It currently
reports isReady immediately, before the provisioned JNLP node has had
a chance to establish its connection with the Jenkins master.

This can result in the OpenStack plugin continually provisioning JNLP
nodes until a JNLP agent establishes a connection to master and reduces
the `excessWorkload` to 0, the OpenStack tenant runs out of compute
resources, or the global instance cap is reached -whichever comes first.

To fix this I've done two things:

1. JCloudsCloud::JNLP::isReady now reports based on the status of the
slave's communication channel.

   Similar to the SSH path, the OpenStack plugin now holds off on
   provisioning new nodes until JNLP agent has established its connection.

2. Add the JNLP slave to Jenkin's list of active connections

   Without this step, if JCloudsCloud::JNLP::isReady returns false and the
   JNLP agent has yet to establish a connection, the JCloudsCleanupThread
   will immediately reclaim the JNLP node in JCloudsCleanupThread::destroyServersOutOfScope
   because the unconnected JNLP node is out of scope i.e. the node is not
   found in the list of active nodes...thus preventing the node from ever
   provisioning.

   Adding the JNLP node to the list of active nodes prevents the
   JDCLoudsCleanup thread from reclaiming the node immediately,
   but I'm unsure if it's the right approach.

   It's modeled after the JCloudsCloud::doProvision step for manually
   provisioning nodes by clicking the button in the computer list. This
   manual step always adds the slave, regardless of type, so I figure this
   is safe from a conceptual point of view, but I'm unsure if
   JDCloudsCloud::JNLP::createLauncher is the most appropriate place for
   this action to occur. I'm also unsure why the SSH agent doesn't also
   require this step.

The end result, is that the OpenStack plugin no longer provisions excess
nodes, and now appropriately waits for JNLP nodes to connect.