Online help improvements.
Help text for EndpointURL (.../JCloudsCloud/help-endPointUrl.html) now
explains how to find out what the URL is likely to be and provides
examples.
Help text for UserDataId (.../SlaveOptions/help-userDataId.jelly, the bit in
the template where the user selects which UserData script to use) now
explains why you need this.
Help text for UserData (.../UserDataConfig/help-userData.jelly, where the user
defines the UserData scripts for use in the templates) no longer says this is
where you select which script to use (this was a copy/paste error from when
the help was relocated).