ProvisioningTest was unreliable.
ProvisioningTest checks Cloud Statistics data.  This data is updated
asynchronously, but the tests did not account for this.  This resulted
in test execution being unpredictable, depending on thread scheduling.
The fix was to add a waitForCloudStatistics method.

It also checks AsyncResourceDisposer effects, and this is also run
asynchronously.
The fix was to add a waitForAsyncResourceDisposer method.