Cache data we get back from Openstack service endpoint.
A number of functions we call will return the same data no matter when
or how we call them, e.g. the list of IP pools does not change often.
It should, therefore, be safe to cache such information and thus avoid
making needless calls to the Openstack service, which in turn will
improve UI responsiveness.
This change caches (for 15 seconds) the results of the following
openstack4j calls:
 - compute().flavors().list()
 - compute().floatingIps().getPoolNames()
 - compute().keypairs().list()
 - images().listAll()
 - images().listAll(name)
 - networking().network().list()