JENKINS-41126: Can now start instances from an OpenStack VolumeSnapshot.
New functionality: Introduced new "bootSource" field that says what kind
of image the imageId is (Image or VolumeSnapshot).  We are no longer
limited to just booting from an Image (github issue #136).
Enhancement: In the configuration panels, drop-down-lists now always
list the currently-chosen option even if it isn't valid anymore, so we
always have the ability to leave things as-is.
Enhancement: validation of the imageId field now checks that the
selected image exists and is non-ambiguous.
Enhancement: validation of AvailabilityZone field added (we now warn if
it is ambiguous).
Enhancement: if selected image is ambiguous, we now use the most recent
match instead of a random one (github issue #119).
Bugfix: Was failing to find images by id if id contained numbers 2-9
(only 0,1 and a-f were accepted).
Improved help text for availabilityZone and image.
Added unit tests for the way we list, lookup and validate images &
volume snapshots.
Refactor: Moved some common Descriptor functionality from
SlaveOptionsDescriptor to new OsAuthDescriptor class.