Add boot from volume with images
This patch allows the creation of volumes by Nova using block device mapping with custom sizes.

This is very useful if you have an instance type with root_gb=0 and you need to control the size of the root volume.

Note that this requires the Cinder block storage service to be active in the cloud.