Add optional instance pre-provisioning
In order to have instances at the ready at all times and
thus speed up CI runs because instance provisioning is done
upfront, add a mechanism and the related parameter to configure
a minimum number of instances that must be ready at all times.

This change is backward-compatible: if the new parameter 'instancesMin'
is set to zero (its default value), no pre-provisioning happens.

By combining a non-zero value of 'instancesMin' and a zero retention
time, one can favour the pre-provisioning of new instances over the
reuse of slaves.

Drive-by: fix test whose failure wasn't checked.