Add metadata to the servers we start that describes the bootsource.
We now enhance the OpenStack server's metadata to describe exactly what
the server was booted from, allowing administrators to track "what
happened" even when names are ambiguous or change over time.
Likewise, in the case of a VolumeSnapshot, when we set the name of the
Volume to specify the VolumeSnapshot it came from, we now also include
the VolumeSnapshot ID.
Also fixed a couple of javadoc warnings by adding missing @return
statements in the file we modified.