Refactor: Now shows live data from OpenStack
Refactor: Moved code out of JCloudsComputer into JCloudsSlave.
Refactor: Renamed openstackMetadata and openstackSlaveOptions to be liveOpenstackServerDetails and openstackSlaveData.
Enhancement: liveOpenstackServerDetails is now "live" in that we'll read the data direct from OpenStack when we need to instead of persisting the data when we first create the slave.  This means that OpenStack-side changes (e.g. turning off or deleting the VM) will be visible from the Jenkins UI.
Enhancement: Gave JCloudsSlave a short-duration cache in which it can store expensive-to-obtain data, so we don't DoS-attack our OpenStack API.
Enhancement: WebUI can now says if the OpenStack data is available or not, so you can tell if someone's deleted the VM.
Enhancement: Now shows lots of OpenStack VM details.