Consider IP addresses without an explicit type as fixed
Issue has been found while provisionning servers with Rackspace.
The given set of addresses looks like that:
```
[
    [NovaAddress{address=10.181.64.154, version=4, }],
    [NovaAddress{address=2a00:1a48:7807:101:ae76:4eff:fe08:487, version=6, }, NovaAddress{address=134.213.24.184, version=4, }]
]
```

Since there is no explicit type, the plugin fails to find a valid IP
and the provisioning of the agent in Jenkins is not successful.

The explicit verification of the type seems to have been introduced with
a refectoring 84900e383bd16de327d9fe3075d4a7d694a49418.