Failing to set name of Volume we booted from is no longer fatal
If we boot from a VolumeSnaphot, the Instance has a Volume attached to
it, but OpenStack doesn't automatically set its name and it sometimes
forgets to remove it with the Instance, so the plugin sets the name to
make it easier for humans to figure out who owns what Volume.
...however this doesn't always work with all versions of OpenStack, and
it's better that we live without this naming than fail to work at all.